% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data_tag.R
\name{add_data_tag}
\alias{add_data_tag}
\title{m2c2R}
\usage{
add_data_tag(data, tag_name = "", tag_value = T)
}
\arguments{
\item{data}{Name of tibble or data frame}

\item{tag_name}{Name of tag to be added to the data, in quotes}

\item{tag_value}{TRUE/FALSE}
}
\value{
Returns input data frame with attributes added.
}
\description{
m2c2R
}
\details{
This function is used by the scoring functions in the m2c2R package
to flag whether data has been scored or not.
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
