% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_metadate_cols.R
\name{append_metadate_cols}
\alias{append_metadate_cols}
\title{m2c2R}
\usage{
append_metadate_cols(
  data,
  time_var = "timestamp",
  time_format = "\%m/\%d/\%Y",
  time_origin = "1970-01-01",
  timezone = Sys.timezone(),
  append_rounded = T
)
}
\description{
m2c2R
}
