% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_n_batches.R
\name{create_n_batches}
\alias{create_n_batches}
\title{m2c2R}
\usage{
create_n_batches(.data, batches = 10)
}
\arguments{
\item{batches}{Number of batches desired}

\item{data}{Name of a tibble or data frame}
}
\value{
Returns list of data, split into specified number ofbatches
}
\description{
m2c2R
}
\details{
This function splits an input data frame into a specified number of batches for processing.
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
