% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_ext.R
\name{get_file_ext}
\alias{get_file_ext}
\title{m2c2R}
\usage{
get_file_ext(filepath)
}
\arguments{
\item{filepath}{Path to a file}
}
\value{
Returns file extension
}
\description{
m2c2R
}
\details{
This function takes a file path as input and returns the file extension
(csv, txt, xls, etc.)
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
