% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_data.R
\name{inspect_data}
\alias{inspect_data}
\title{m2c2R}
\usage{
inspect_data(data)
}
\arguments{
\item{data}{Name of a tibble or data frame}
}
\description{
m2c2R
}
\details{
This function "inspects" the data by pulling the number of rows and 
columns, as well as the variable names.
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
