% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_col_present.R
\name{is_col_present}
\alias{is_col_present}
\title{m2c2R}
\usage{
is_col_present(data, colname)
}
\arguments{
\item{data}{Name of a tibble or data frame}

\item{colname}{Name of a column to check for, in quotes ("participant_id")}
}
\value{
Returns TRUE/FALSE.
}
\description{
m2c2R
}
\details{
This function checks whether a given column (colname) is in the 
specified data frame or tibble.
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
