% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_data_frame_tibble.R
\name{is_data_frame_tibble}
\alias{is_data_frame_tibble}
\title{m2c2R}
\usage{
is_data_frame_tibble(data)
}
\arguments{
\item{data}{Name of an object in R}
}
\value{
Returns TRUE/FALSE.
}
\description{
m2c2R
}
\details{
This function checks whether a given object is a data frame or tibble.
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
