% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tidy_datetime.R
\name{make_tidy_datetime}
\alias{make_tidy_datetime}
\title{m2c2R}
\usage{
make_tidy_datetime(delim = "_")
}
\arguments{
\item{delim}{Delimiter to use in separating parts of the datetime object.}
}
\value{
Returns formatted datetime.
}
\description{
m2c2R
}
\details{
This function pulls and formats the Sys.time() from the current R session, using
the specified delimiter.
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
