% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_log_summary_export.R
\name{produce_log_summary_export}
\alias{produce_log_summary_export}
\title{m2c2R}
\usage{
produce_log_summary_export(
  filepath,
  part_ids = NA,
  search_events = NA,
  debug = F
)
}
\arguments{
\item{filepath}{Path to a file}
}
\value{
Returns input data frame with m2c2_processing_hash and m2c2_processing_timestamp added.
}
\description{
m2c2R
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
