% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcsv.R
\name{qcsv}
\alias{qcsv}
\title{m2c2R}
\usage{
qcsv(., ts = T, tz = "UTC")
}
\arguments{
\item{.}{class: data.frame}

\item{ts}{class: boolean}

\item{tz}{class: string}
}
\description{
m2c2R
}
\examples{
qcsv(., ts=T, tz="UTC")
}
