% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_stroop.R
\name{score_stroop}
\alias{score_stroop}
\title{m2c2R}
\usage{
score_stroop(data)
}
\arguments{
\item{data}{name of tibble or data frame}
}
\value{
Returns a data frame of scored data
}
\description{
m2c2R
}
\details{
Scores Stroop/Color Naming Task data. Takes tibble or data frame as
input, renames necessary variables, adds data tag to mark the data as 
scored, and adds columns to denote the scoring (m2c2_processing_hash, 
m2c2_processing_timestamp).
}
