% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stroop.R
\name{summary_stroop}
\alias{summary_stroop}
\title{m2c2R}
\usage{
summary_stroop(
  data,
  group_var,
  var_prefix = "stroop",
  experimental = F,
  lure_label = "LURE",
  normal_label = "NORMAL"
)
}
\description{
m2c2R
}
